Syntax secondaries generated, sub-headings included in secondary, escaping fixed

+++ plugins/html/secondaries_v0_auth_collection/cleanup.raku
@@ -8,6 +8,6 @@ sub ($pr, %processed, %options --> Array) {
             .unlink
         }
     }
-    'prettyurls'.IO.unlink;
+    'prettyurls'.IO.unlink if 'prettyurls'.IO ~~ :e & :f;
     []
 }
